requirements.txt
fastapi
uvicorn[standard]
Pandas 
transformers
typing
pillow
matplotlib
qdrant-client